requirements.txt
onnxtr[cpu]
qqdm
PyYAML
tqdm
pandas
numpy
opencv-python
Pillow
onnxruntime 
matplotlib
gradio-image-prompter
gradio
torch
tensorboard
onnx
torchsummaryX
python-docx
scipy